requirements.txt
git+https://github.com/nghorbani/configer
git+https://github.com/mattloper/chumpy
setuptools==59.5.0
torchgeometry==0.1.2
tensorboard
numpy
smplx
pyrender
open3d
imageio-ffmpeg
matplotlib
opencv-python
scipy
scikit-image
joblib
cython
tqdm
hydra-core
pyyaml
gdown
dill
motmetrics
einops
mmcv==1.3.9
timm==0.4.9
xtcocotools
pandas==1.4.0